split change canister from change external canister
