+++ core/station/impl/src/factories/requests/change_canister.rs
@@ -2,8 +2,8 @@ use super::{Create, Execute, RequestExecuteStage};
 use crate::{
     errors::{RequestError, RequestExecuteError},
     models::{
-        ChangeCanisterOperation, ChangeCanisterTarget, ChangeExternalCanisterOperation, Request,
-        RequestExecutionPlan, RequestOperation,
+        ChangeCanisterOperation, ChangeCanisterTarget, Request, RequestExecutionPlan,
+        RequestOperation,
     },
     services::{ChangeCanisterService, DisasterRecoveryService, SystemService},
 };
@@ -11,9 +11,7 @@ use async_trait::async_trait;
 use candid::Encode;
 use orbit_essentials::types::UUID;
 use sha2::{Digest, Sha256};
-use station_api::{
-    ChangeCanisterOperationInput, ChangeExternalCanisterOperationInput, CreateRequestInput,
-};
+use station_api::{ChangeCanisterOperationInput, CreateRequestInput};
 use std::sync::Arc;
 
 pub struct ChangeCanisterRequestCreate;
@@ -137,91 +135,5 @@ impl Execute for ChangeCanisterRequestExecute<'_, '_> {
                 Ok(RequestExecuteStage::Completed(operation))
             }
         }
-    }
-}
-
-pub struct ChangeExternalCanisterRequestCreate;
-
-#[async_trait]
-impl Create<ChangeExternalCanisterOperationInput> for ChangeExternalCanisterRequestCreate {
-    async fn create(
-        &self,
-        request_id: UUID,
-        requested_by_user: UUID,
-        input: CreateRequestInput,
-        operation_input: ChangeExternalCanisterOperationInput,
-    ) -> Result<Request, RequestError> {
-        let request = Request::new(
-            request_id,
-            requested_by_user,
-            Request::default_expiration_dt_ns(),
-            RequestOperation::ChangeExternalCanister(ChangeExternalCanisterOperation {
-                arg_checksum: operation_input.arg.as_ref().map(|arg| {
-                    let mut hasher = Sha256::new();
-                    hasher.update(arg);
-                    hasher.finalize().to_vec()
-                }),
-                module_checksum: {
-                    let mut hasher = Sha256::new();
-                    hasher.update(&operation_input.module);
-                    hasher.finalize().to_vec()
-                },
-                input: operation_input.into(),
-            }),
-            input
-                .execution_plan
-                .map(Into::into)
-                .unwrap_or(RequestExecutionPlan::Immediate),
-            input
-                .title
-                .unwrap_or_else(|| "ChangeExternalCanister".to_string()),
-            input.summary,
-        );
-
-        Ok(request)
-    }
-}
-
-pub struct ChangeExternalCanisterRequestExecute<'p, 'o> {
-    request: &'p Request,
-    operation: &'o ChangeExternalCanisterOperation,
-    change_canister_service: Arc<ChangeCanisterService>,
-}
-
-impl<'p, 'o> ChangeExternalCanisterRequestExecute<'p, 'o> {
-    pub fn new(
-        request: &'p Request,
-        operation: &'o ChangeExternalCanisterOperation,
-        change_canister_service: Arc<ChangeCanisterService>,
-    ) -> Self {
-        Self {
-            request,
-            operation,
-            change_canister_service,
-        }
-    }
-}
-
-#[async_trait]
-impl Execute for ChangeExternalCanisterRequestExecute<'_, '_> {
-    async fn execute(&self) -> Result<RequestExecuteStage, RequestExecuteError> {
-        self.change_canister_service
-            .install_canister(
-                self.operation.input.canister_id,
-                self.operation.input.mode.clone(),
-                &self.operation.input.module,
-                self.operation.input.arg.clone(),
-            )
-            .await
-            .map_err(|err| RequestExecuteError::Failed {
-                reason: format!(
-                    "failed to install external canister {}: {}",
-                    self.operation.input.canister_id, err
-                ),
-            })?;
-
-        Ok(RequestExecuteStage::Completed(
-            self.request.operation.clone(),
-        ))
     }
 }
